+++ src/components/index.js
@@ -3,5 +3,13 @@ import LoadingModal from './LoadingModal';
 import Menlo from './Menlo';
 import NotificationModal from './NotificationModal';
 import NotificationRow from './NotificationRow';
+import SettingRow from './SettingRow';
 
-export {AlertModal, LoadingModal, Menlo, NotificationModal, NotificationRow};
+export {
+  AlertModal,
+  LoadingModal,
+  Menlo,
+  NotificationModal,
+  NotificationRow,
+  SettingRow,
+};