+++ src/views/index.js
@@ -1,7 +1,10 @@
+import About from './About';
+import Api from './Api';
 import Changelog from './Changelog';
+import Help from './Help';
 import Home from './Home';
 import License from './License';
 import Login from './Login';
 import Settings from './Settings';
 
-export {Changelog, Home, License, Login, Settings};
+export {About, Api, Changelog, Help, Home, License, Login, Settings};